UIBULKED-106: refactor the code
